requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-stubs==5.0.0
django-stubs-ext==5.0.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.7
mysql-connector-python==8.4.0
mysqlclient==2.2.4
oauthlib==3.2.2
phonenumbers==8.13.36
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
